Skille ny periode ut i egen komponent
